requirements.txt
flask>=2.3.0
requests>=2.31.0
openai>=1.0.0
python-dotenv>=1.0.0
schedule>=1.2.0
py-clob-client>=0.0.1
